Implement better line moving by keeping track of original vertex positions
Line moves are now implemented by keeping track of the original
position of the two vertices surrounding the moved line. This
makes the line move follow the cursor more smoothly.
There are still issues with how vertices are affected by moving
a line (i.e. if a line is moved beyond a surrounding vertex, it
get "squished" to the line; if the line is moved vertically far
enough that one of the affected vertices hits the top or bottom
of the widget, the other vertex is able to keep being moved)